refactor(communitybanner): make CardView look nice with custom Tabs
