+++ src/components/TabBar/DesktopView/NormalView.js
