+++ src/components/TabBar/styles/desktop_view/brief_view.js
